Fixed read from access->get() data

+++ source/FAST/Algorithms/UsReconstruction/Us3Dhybrid/Us3Dhybrid.hpp
@@ -103,7 +103,7 @@ namespace fast {
         int zLocMultiplier;
         ImageAccess::pointer frameAccess;
         Vector3ui frameSize;
-        int * frameData;
+        uchar * frameData;
         uint frameChannels;
         DataType frameType;
